+++ src/aks-preview/azext_aks_preview/_format.py
@@ -70,6 +70,23 @@ def aks_machine_show_table_format(result):
     return parser(result)
 
 
+def aks_operation_show_table_format(result):
+    def parser(entry):
+        percentComplete = ""
+        if entry["percentComplete"]:
+            percentComplete = str(entry["percentComplete"]) + "%"
+        entry["percentComplete"] = percentComplete
+        parsed = compile_jmes("""{
+                name: name,
+                status: status,
+                startTime: startTime,
+                endTime: endTime,
+                percentComplete: percentComplete
+            }""")
+        return parsed.search(entry, Options(dict_cls=OrderedDict))
+    return parser(result)
+
+
 def aks_agentpool_show_table_format(result):
     """Format an agent pool as summary results for display with "-o table"."""
     return [_aks_agentpool_table_format(result)]