+++ src/containerapp/azext_containerapp/containerapp_sessionpool_decorator.py
@@ -171,7 +171,8 @@ class SessionPoolCreateDecorator(SessionPoolPreviewDecorator):
 
             container_def = self.set_up_container()
             ingress_def = self.set_up_ingress()
-            registry_def = self.set_up_registry_auth_configuration(secrets_def)
+            registry_def, updated_secret_def = self.set_up_registry_auth_configuration(secrets_def)
+            secrets_def = updated_secret_def
 
             customer_container_template["containers"] = [container_def]
             customer_container_template["ingress"] = ingress_def
@@ -218,7 +219,7 @@ class SessionPoolCreateDecorator(SessionPoolPreviewDecorator):
 
     def set_up_container(self):
         container_def = ContainerModel
-        container_def["name"] = self.get_argument_container_name() if self.get_argument_container_name() else self.get_argument_name()
+        container_def["name"] = self.get_argument_container_name() if self.get_argument_container_name() else self.get_argument_name().lower()
         container_def["image"] = self.get_argument_image() if self.get_argument_image() else HELLO_WORLD_IMAGE
         if self.get_argument_env_vars() is not None:
             container_def["env"] = parse_env_var_flags(self.get_argument_env_vars())
@@ -239,7 +240,7 @@ class SessionPoolCreateDecorator(SessionPoolPreviewDecorator):
         registry_def = None
         if self.get_argument_registry_server() is not None:
             registry_def = {}
-            registry_def["registryServer"] = self.get_argument_registry_server()
+            registry_def["server"] = self.get_argument_registry_server()
             registry_def["username"] = self.get_argument_registry_user()
 
             if secrets_def is None:
@@ -248,7 +249,7 @@ class SessionPoolCreateDecorator(SessionPoolPreviewDecorator):
                                                                                       self.get_argument_registry_user(),
                                                                                       self.get_argument_registry_server(),
                                                                                       self.get_argument_registry_pass())
-        return registry_def
+        return registry_def, secrets_def
 
     def set_up_ingress(self):
         if self.get_argument_target_port() is None:
@@ -415,17 +416,31 @@ class SessionPoolUpdateDecorator(SessionPoolPreviewDecorator):
 
     def set_up_registry_auth_configuration(self, secrets_def, customer_container_template):
         if self.get_argument_registry_server() is not None:
-            safe_set(customer_container_template, "registryCredentials", "registryServer", value=self.get_argument_registry_server())
+            safe_set(customer_container_template, "registryCredentials", "server", value=self.get_argument_registry_server())
         if self.get_argument_registry_user() is not None:
             safe_set(customer_container_template, "registryCredentials", "username", value=self.get_argument_registry_user())
         if secrets_def is None:
             secrets_def = []
         if self.get_argument_registry_pass() is not None:
+            original_secrets = self.existing_pool_def["properties"]["secrets"]
+            original_secrets_names = []
+            for secret in original_secrets:
+                original_secrets_names.append(secret["name"])
             safe_set(customer_container_template, "registryCredentials", "passwordSecretRef",
                      value=store_as_secret_and_return_secret_ref(secrets_def,
-                                                                 self.get_argument_registry_user(),
-                                                                 self.get_argument_registry_server(),
+                                                                 customer_container_template["registryCredentials"]["username"],
+                                                                 customer_container_template["registryCredentials"]["server"],
                                                                  self.get_argument_registry_pass()))
+            new_secret_names = []
+            for secret in secrets_def:
+                new_secret_names.append(secret["name"])
+            deleted_secrets = set(original_secrets_names).difference(new_secret_names)
+            if len(deleted_secrets) > 0:
+                logger.warning("the following secrets are going to be deleted: " + str(deleted_secrets) + " If this is not the intended behavior, please add the missing secrets into the --secrets flag.")
+
+            # Update the secrets to the patch payload.
+            if len(secrets_def) > 0:
+                safe_set(self.session_pool_def, "properties", "secrets", value=secrets_def)
 
     def set_up_ingress(self, customer_container_template):
         if self.get_argument_target_port() is not None: