+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -168,6 +168,41 @@ class ContainerappIdentityTests(ScenarioTest):
             JMESPathCheck('type', 'None'),
         ])
 
+    @AllowLargeResponse(8192)
+    @ResourceGroupPreparer(location="westeurope")
+    def test_containerapp_identity_keda(self, resource_group):
+        # MSI is not available in North Central US (Stage), if the TEST_LOCATION is "northcentralusstage", use eastus as location
+        location = TEST_LOCATION
+        if format_location(location) == format_location(STAGE_LOCATION):
+            location = "eastus"
+        self.cmd('configure --defaults location={}'.format(location))
+
+        ca_name = self.create_random_name(prefix='containerapp', length=24)
+        user_identity_name1 = self.create_random_name(prefix='containerapp-user1', length=24)
+
+        env = prepare_containerapp_env_for_app_e2e_tests(self, location=location)
+
+        user_identity_id = self.cmd('identity create -g {} -n {}'.format(resource_group, user_identity_name1)).get_output_in_json()["id"]
+
+        self.cmd(f'containerapp create -g {resource_group} -n {ca_name} --environment {env} --system-assigned --user-assigned {user_identity_name1} --scale-rule-name azure-queue --scale-rule-type azure-queue --scale-rule-metadata "accountName=account1" "queueName=queue1" "queueLength=1" --scale-rule-identity {user_identity_name1}')
+        self.cmd(f'containerapp show -g {resource_group} -n {ca_name}', checks=[
+            JMESPathCheck("properties.template.scale.rules[0].name", "azure-queue"),
+            JMESPathCheck("properties.template.scale.rules[0].azureQueue.accountName", "account1"),
+            JMESPathCheck("properties.template.scale.rules[0].azureQueue.queueName", "queue1"),
+            JMESPathCheck("properties.template.scale.rules[0].azureQueue.queueLength", "1"),
+            JMESPathCheck("properties.template.scale.rules[0].azureQueue.identity", user_identity_id, case_sensitive=False),
+        ])
+
+        self.cmd(f'containerapp update -g {resource_group} -n {ca_name} --scale-rule-name azure-blob --scale-rule-type azure-blob --scale-rule-metadata "accountName=account2" "blobContainerName=blob2" "blobCount=2" --scale-rule-identity {user_identity_id}')
+        self.cmd(f'containerapp show -g {resource_group} -n {ca_name}', checks=[
+            JMESPathCheck("properties.template.scale.rules[0].name", "azure-blob"),
+            JMESPathCheck("properties.template.scale.rules[0].custom.metadata.accountName", "account2"),
+            JMESPathCheck("properties.template.scale.rules[0].custom.metadata.blobContainerName", "blob2"),
+            JMESPathCheck("properties.template.scale.rules[0].custom.metadata.blobCount", "2"),
+            JMESPathCheck("properties.template.scale.rules[0].custom.identity", user_identity_id, case_sensitive=False),
+            JMESPathCheck("properties.template.scale.rules[0].custom.type", "azure-blob"),
+        ])
+
 
 class ContainerappIngressTests(ScenarioTest):
     def __init__(self, *arg, **kwargs):
@@ -2501,12 +2536,6 @@ class ContainerappRuntimeTests(ScenarioTest):
         # Create container app with runtime=java, it should have default java runtime settings
         create_containerapp_with_runtime_java_metrics_args_and_check('--runtime=java', checks=[
                 JMESPathCheck('properties.provisioningState', "Succeeded"),
-                JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False)
-            ])
-
-        # Create container app with runtime=java and enable java metrics
-        create_containerapp_with_runtime_java_metrics_args_and_check('--runtime=java --enable-java-metrics', checks=[
-                JMESPathCheck('properties.provisioningState', "Succeeded"),
                 JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True)
             ])
 
@@ -2514,6 +2543,12 @@ class ContainerappRuntimeTests(ScenarioTest):
         create_containerapp_with_runtime_java_metrics_args_and_check('--runtime=java --enable-java-metrics=false', checks=[
                 JMESPathCheck('properties.provisioningState', "Succeeded"),
                 JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False)
+            ])
+
+        # Create container app with runtime=java and enable java metrics
+        create_containerapp_with_runtime_java_metrics_args_and_check('--runtime=java --enable-java-metrics', checks=[
+                JMESPathCheck('properties.provisioningState', "Succeeded"),
+                JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True)
             ])
 
         # Create container app with enable java metrics without setting runtime explicitly
@@ -2567,7 +2602,13 @@ class ContainerappRuntimeTests(ScenarioTest):
         # Update container app with runtime=java, it should setup default java runtime settings if not set before
         self.cmd(f'containerapp update -g {resource_group} -n {app} --runtime=java', checks=[
                 JMESPathCheck('properties.provisioningState', "Succeeded"),
-                JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False)
+                JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True)
+            ])
+
+        # Update container app with only runtime=java will keep the previous settings if set before
+        self.cmd(f'containerapp update -g {resource_group} -n {app} --runtime=java', checks=[
+                JMESPathCheck('properties.provisioningState', "Succeeded"),
+                JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True)
             ])
 
         self.cmd(f'containerapp update -g {resource_group} -n {app} --runtime=java --enable-java-metrics', checks=[
@@ -2586,6 +2627,7 @@ class ContainerappRuntimeTests(ScenarioTest):
                 JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False)
             ])
 
+        # Update container app with only runtime=java will keep the previous settings if set before
         self.cmd(f'containerapp update -g {resource_group} -n {app} --runtime=java', checks=[
                 JMESPathCheck('properties.provisioningState', "Succeeded"),
                 JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False)
@@ -2636,13 +2678,13 @@ class ContainerappRuntimeTests(ScenarioTest):
         # Create container app with runtime=java, it should have default java runtime settings
         create_containerapp_with_runtime_java_agent_args_and_check('--runtime=java', checks=[
             JMESPathCheck('properties.provisioningState', "Succeeded"),
-            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False),
+            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True),
         ])
 
         # Create container app with runtime=java and enable java agent
         create_containerapp_with_runtime_java_agent_args_and_check('--runtime=java --enable-java-agent', checks=[
             JMESPathCheck('properties.provisioningState', "Succeeded"),
-            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False),
+            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True),
             JMESPathCheck("properties.configuration.runtime.java.javaAgent.enabled", True)
         ])
 
@@ -2763,7 +2805,7 @@ class ContainerappRuntimeTests(ScenarioTest):
         # Update container app with runtime=java, it should setup default java runtime settings if not set before
         self.cmd(f'containerapp update -g {resource_group} -n {app} --runtime=java', checks=[
             JMESPathCheck('properties.provisioningState', "Succeeded"),
-            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", False),
+            JMESPathCheck("properties.configuration.runtime.java.enableMetrics", True),
             JMESPathCheckNotExists("properties.configuration.runtime.java.javaAgent")
         ])