+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/devcenter/_show.py
@@ -22,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-10-01-preview",
+        "version": "2024-05-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}", "2023-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}", "2024-05-01-preview"],
         ]
     }
 
@@ -101,7 +101,7 @@ class Show(AAZCommand):
 
         @property
         def error_format(self):
-            return "ODataV4Format"
+            return "MgmtErrorFormat"
 
         @property
         def url_parameters(self):
@@ -125,7 +125,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-10-01-preview",
+                    "api-version", "2024-05-01-preview",
                     required=True,
                 ),
             }
@@ -210,6 +210,9 @@ class Show(AAZCommand):
             )
 
             properties = cls._schema_on_200.properties
+            properties.dev_box_provisioning_settings = AAZObjectType(
+                serialized_name="devBoxProvisioningSettings",
+            )
             properties.dev_center_uri = AAZStrType(
                 serialized_name="devCenterUri",
                 flags={"read_only": True},
@@ -218,9 +221,23 @@ class Show(AAZCommand):
                 serialized_name="displayName",
             )
             properties.encryption = AAZObjectType()
+            properties.network_settings = AAZObjectType(
+                serialized_name="networkSettings",
+            )
+            properties.plan_id = AAZStrType(
+                serialized_name="planId",
+            )
+            properties.project_catalog_settings = AAZObjectType(
+                serialized_name="projectCatalogSettings",
+            )
             properties.provisioning_state = AAZStrType(
                 serialized_name="provisioningState",
                 flags={"read_only": True},
+            )
+
+            dev_box_provisioning_settings = cls._schema_on_200.properties.dev_box_provisioning_settings
+            dev_box_provisioning_settings.install_azure_monitor_agent_enable_status = AAZStrType(
+                serialized_name="installAzureMonitorAgentEnableStatus",
             )
 
             encryption = cls._schema_on_200.properties.encryption
@@ -245,6 +262,16 @@ class Show(AAZCommand):
             )
             key_encryption_key_identity.user_assigned_identity_resource_id = AAZStrType(
                 serialized_name="userAssignedIdentityResourceId",
+            )
+
+            network_settings = cls._schema_on_200.properties.network_settings
+            network_settings.microsoft_hosted_network_enable_status = AAZStrType(
+                serialized_name="microsoftHostedNetworkEnableStatus",
+            )
+
+            project_catalog_settings = cls._schema_on_200.properties.project_catalog_settings
+            project_catalog_settings.catalog_item_sync_enable_status = AAZStrType(
+                serialized_name="catalogItemSyncEnableStatus",
             )
 
             system_data = cls._schema_on_200.system_data